ECO-174: tutorial fixes.
