+++ .gitignore
@@ -1,2 +1,3 @@
 target
 Cargo.lock
+tests/wasm/contract.wasm
