+++ contract/src/env.rs
@@ -1,4 +1,4 @@
-use core::convert::TryInto;
+use std::convert::TryInto;
 
 use crate::{
     api::{self},
