+++ contract/src/lib.rs
@@ -1,8 +1,3 @@
-#![no_std]
-
-#[macro_use]
-extern crate alloc;
-
 mod api;
 mod contracts;
 mod env;
