+++ logic/Cargo.toml
@@ -1,16 +1,13 @@
 [package]
-name = "erc20-logic"
+name = "logic"
 version = "0.1.1"
 authors = ["Maciej Zieliński <maciej@casperlabs.io>"]
 edition = "2018"
 
 [lib]
-name = "erc20_logic"
+name = "logic"
 doctest = false
-
-[features]
-default = ["std"]
-std = []
+bench = false
 
 [dependencies]
 num-traits = { version = "0.2.10", default-features = false }
